Use uint64_t instead of size_t for table, since size_t is 32-bit on 32-bit systems.

+++ include/matcher.h
@@ -10,10 +10,10 @@
 #include "util.h"
 
 struct SFHASH_FileMatcher {
-  std::vector<std::pair<size_t, sha1_t>> table;
+  std::vector<std::pair<uint64_t, sha1_t>> table;
   std::unique_ptr<ProgramHandle, decltype(lg_destroy_program)&> prog;
 };
 
-std::tuple<std::string, size_t, sha1_t> parse_line(const char* beg, const char* const end);
+std::tuple<std::string, uint64_t, sha1_t> parse_line(const char* beg, const char* const end);
 
 std::unique_ptr<SFHASH_FileMatcher> load_hashset(const char* beg, const char* end, LG_Error** err);
